+++ packages/suite/src/utils/wallet/sendFormUtils.ts
@@ -1,9 +1,16 @@
-import { Output, State, EthTransactionData, EthPreparedTransaction } from '@wallet-types/sendForm';
+import {
+    Output,
+    State,
+    EthTransactionData,
+    EthPreparedTransaction,
+    FeeInfo,
+    FeeLevel,
+} from '@wallet-types/sendForm';
 import { Account } from '@wallet-types';
 import { VALIDATION_ERRORS } from '@wallet-constants/sendForm';
 import BigNumber from 'bignumber.js';
 import { formatNetworkAmount } from '@wallet-utils/accountUtils';
-import { toHex, toWei } from 'web3-utils';
+import { toHex, toWei, fromWei } from 'web3-utils';
 import { Transaction } from 'ethereumjs-tx';
 
 export const getOutput = (outputs: Output[], id: number) =>
@@ -164,4 +171,27 @@ export const getReserveInXrp = (account: Account) => {
     if (account.networkType !== 'ripple') return null;
     const { misc } = account;
     return formatNetworkAmount(misc.reserve, account.symbol);
+};
+
+export const getFeeLevels = (account: Account, feeInfo: FeeInfo) => {
+    const convertedEthLevels: FeeLevel[] = [];
+    const initialLevels: FeeLevel[] =
+        account.networkType === 'ethereum'
+            ? feeInfo.levels
+            : feeInfo.levels.concat({
+                  label: 'custom',
+                  feePerUnit: '0',
+                  blocks: -1,
+              });
+
+    if (account.networkType === 'ethereum') {
+        initialLevels.forEach(level =>
+            convertedEthLevels.push({
+                ...level,
+                feePerUnit: fromWei(level.feePerUnit, 'gwei'),
+            }),
+        );
+    }
+
+    return account.networkType === 'ethereum' ? convertedEthLevels : initialLevels;
 };
